Repackages the in-tree Notion (8A) + Obsidian (8B) connectors as opt-in
sandboxed plugins on the 9A provides.issueSource bridge, wires the host-side
poll loop 9A left inert (replaces Obsidian's un-sandboxable chokidar watcher),
and extends the coexistence yield to Process B. Combined phase + full Obsidian
writeback (Chris-approved). Opus audit 0C/1M/5m all resolved; 5 acceptance
gates green incl. a real-bundle smoke that caught a gray-matter CJS-under-ESM
crash. 9B COMPLETE; 9C next.

+++ src/plugins/issue-connector-adapter.ts
@@ -72,7 +72,7 @@ const NormalizedIssueSchema = z.object({
   assignee: z.string().nullable().catch(null).default(null),
   labels: z.array(z.string()).catch([]).default([]),
   projectValue: z.string().nullable().catch(null).default(null),
-  priority: z.number().catch(0).default(0),
+  priority: z.number().int().catch(0).default(0),
   updatedAt: z.string().nullable().catch(null).default(null),
 });
